CCM-1031: Restructure tests
CCM-1031: Restructure tests

CCM-1031: Restructure tests

CCM-1031: Restructure tests

CCM-1031: Restructure tests

CCM-1031: Restructure tests
