+++ tests/sandbox/test_message_batch_create.py
@@ -1,6 +1,3 @@
-"""
-Server returns a 201 when a valid request is raised against /v1/message-batches
-"""
 import requests
 import pytest
 import string
@@ -34,8 +31,8 @@ def __assert_201_response(resp):
 @pytest.mark.parametrize('accept_headers', VALID_ACCEPT_HEADERS)
 def test_201_message_batch_valid_accept_headers(nhsd_apim_proxy_url, accept_headers):
     resp = requests.post(f"{nhsd_apim_proxy_url}{REQUEST_PATH}", headers={
-            "Accept": accept_headers,
-            "Content-Type": "application/json"
+        "Accept": accept_headers,
+        "Content-Type": "application/json"
         }, json={
             "data": {
                 "type": "MessageBatch",
